started new movegenerator logic approach
